Fix bug after utils file refactoring
