+++ src/components/trip-point-edit.js
@@ -1,6 +1,6 @@
 import AbstractComponent from './abstract-component';
 import {EventTypeEnum, LOCATIONS} from '../mock/consts';
-import {moment} from '../utils';
+import {moment} from '../utils/common';
 import {getEventPlaceholder} from '../mock/trip-event';
 
 /**
